Don't update the tag list when the user is typing
Reset the pending update delay when we get a Scintilla change not to
reparse the tags when the user is actively typing.

This should not be a problem since the user is unlikely to care about
the tags being updated if he's actively typing, and should avoid
annoying delays when actively typing inside very large files.

Patch by Yura Siamashka, thanks.

git-svn-id: https://geany.svn.sourceforge.net/svnroot/geany/trunk@5735 ea778897-0a13-0410-b9d1-a72fbfd435f5