Remove widget reparenting in Split Window plugin
Instead of reparenting the documents notebook full of
ScintillaObjects, just ref it, remove it from the old parent, add
it to the new parent, and then unref it.  This fixes the display
issue on Windows and seems to have no issues on Linux.

Patch by Matthew Brush, thanks.
Closes #2725342.

Re-apply this patch now the X PRIMARY selection bug is fixed in
Scintilla's side (see r5692).

git-svn-id: https://geany.svn.sourceforge.net/svnroot/geany/trunk@5774 ea778897-0a13-0410-b9d1-a72fbfd435f5