Only include config.h when building Geany
config.h is not distributed, and we includ it from geany.h.  Even
though it was only included it HAVE_CONFIG_H was defined, an Autotools
based build system is likely to define it, and it becomes a problem if
that very build system uses another header name than config.h.

Closes #3384026

git-svn-id: https://geany.svn.sourceforge.net/svnroot/geany/trunk@5882 ea778897-0a13-0410-b9d1-a72fbfd435f5