Reload document configuration on tab page switch
This ensures that the type keywords for the current editor get re-updated
to properly highlight when going between two languages which are using
the same lexer (ex. C/C++/Vala).

Call to sidebar_update_tag_list() removed since it seems to be called in
a round-about way from document_reload_config().