Alter warning fix
Uninitialised GValue does not always work, but is an opaque type so
structure of initialiser isn't known.  Glib 2.30 on has G_VALUE_INIT
to use as initial value.  Fix so if not defined give it the previous literal
value { 0 }, although this leaves the warning, so it is not initialising
the GValue correctly but enough to work.