Prevent double free in filetypes_detect_from_file_internal().
Some glib versions (e.g. 2.28.8) don't actually create a GMatchInfo
in g_regex_match() if it retuns false (the documentation states it does).
This happens when the string is NULL and throws a
"g_regex_match_full: assertion `string != NULL' failed"
This leads to a double free since the match_info paramter is unchanged,
and passed to g_match_info_free() (as according to the docs).

Make sure to not pass NULL as happens when filetypes_detect_from_file_internal()
is called from filetypes_detect_from_file().