Make sure not to emit activation/change signals twice from entry action
If connect_proxy() get called twice, then the handlers would have been
connected twice, leading to wrongly emit activation/change signals twice
for a single input signal.

For whatever reason connect_proxy() is actually called twice for the
quick search entry on Windows (but not on Linux), so this fixes the
search entry behavior on Windows.