Drop 'already' from the message in project close confirmation dialog
Suppose you have project A open and want to open project B. Then the message
saying "The 'A' project is already open" displays. This is slightly confusing
and feels like if you were trying to re-open project A even though you
are opening different project. The message without 'already' looks clearer
in this context.