GeanyWrapLabel: make sure we ask for the height matching our width
With this patch, we make sure the height we query is the one matching
the width we stored and not a newer or older one.

This only fixes the sizing problem, the rendering issue visible in
some situations is still present.

This fix is replicated from the fix in libview's WrapLabel, see
http://view.svn.sourceforge.net/viewvc/view/trunk/libview/libview/wrapLabel.cc?revision=132&view=markup
It also includes the "fix for the fix",
http://view.svn.sourceforge.net/viewvc/view/trunk/libview/libview/wrapLabel.cc?revision=135&view=markup