Remove sentence from documentation that doesn't apply anymore
With newer GTK+ releases, leaving this field blank causes the "Recent
Files" "feature" to be used rather the the expected current working
directory. No replacement text added to explain the new behaviour since
it would be incorrect with older GTK+ versions, would likely be incorrect
on Windows with native dialogs and it will possibly be reverted in GTK+
in the future.

Recenerate HTML manual, including previous changes from other commits.