splitwindow: Rework the design to swap editor and view on focus
The second Scintilla window created by the splitwindow plugin cannot
benefit of the features of the original (Geany's) one because of
Geany's design.  The plugin then used to provide an editor with lower
features, making it more a viewer than a second editor widget.

This commit changes the logic so the plugin-created Scintilla widget
never gets the focus but is rather swapped on the fly with the real
Geany editor when focus enters.  This makes Geany's editor the focused
one while keeping the look and feel of split window.

Although this works quite well, there still is a few things that needs
to be polished, especially reflecting selection and folding states upon
swapping.