CoffeeScript Lexer from Scintilla setup
These changes are enabling the CoffeeScript lexer from the Scintilla projekt.
The lexer is based on C and thus Geany can use the highlighting_styles_C for
COFFEESCRIPT, too. The filetypes.coffeescript inherits from C too, which is
working, but not that beauty. Some could adapt the the colours.

CTags are not enabled yet