Allow user gtkrc to override our gtkrc styles
This allows the user to override the custom styles we apply to some
widgets, like e.g. the unmatched search entry colors.

We use the :theme priority rather than the :application one because
it seems that the :application one cannot override theme settings, even
if it matches against a name the theme don't have rules for but have
rules for the class of that widget.  This prevents a theme from
overriding our styles, but it's unlikely a theme actually provide some
specific stuff for us anyway.