Fix GTK default keybindings being still active after being unbound
If we provide an AccelGroup when creating a menu item using a sock ID,
it installs the GTK default accelerator, accelerator we can't remove
since we don't know about it.  So, don't give an AccelGroup so GTK
don't install it's own accelerator.

This fix also required to properly update the accelerator on some item
we used to ignore since the update didn't work anyway (since the GTK
accelerator was displayed instead).

Note that this doesn't fix the fact the editor popup menu accelerators
are never updated after startup so they don't get updated before
restart after changing a keybinding in the preferences.  This is a
separate (and less problematic) issue due to a simple lack of update.

Closes #1912683 and #3599251.