Add Rename document feature
Accessible from MainMenu->Document->Rename, Documents sidebar tab
context menu and tab label context menu. It opens a little dialog
that lets you edit the file's basename. It's meant to be a
simple and quick alternative to using the "Rename" button in the
Save As dialog.

When the dialog is shown, the part of the basename in the entry
before the last period/dot is selected. If one presses Enter on
the entry it invokes the dialog's default response, and if the
basename didn't change, the document is just saved with its old
name (if needed). If the document is not saved on disk yet, a
question dialog is presented notifying the user that an unsaved
document cannot be renamed and offers them the option of saving
the document.

TODO/Notes/Future:
* Should it have a keybinding? Which one by default?
* Which accelerator label underscore thing to use (if any) for the
new menu item labels?
* It might be useful to show a path relative to the project base
dir if a project is open in the rename dialog, for example if
renaming a file from "src/blah.h" to "include/blah.h".
* It might also be useful to put "Clone" into the Documents sidebar
tab and tab label context menus.