Simplify Rename feature by re-using existing Save As dialog
Rather than presenting a new dialog, just launch the normal
"Save As" dialog but ensure the default response button is
set to the "Rename" button. This has the same distance in terms
of clicks/keystrokes but reduces the code complexity by quite
a bit.

Thanks to Colomban and Lex via IRC for the suggestions.