Improve scrolling to display a plugin's keybindings
When scrolling the keybinding list to display a particular row, which
is used to display a particular plugin's keybindings, consistently
scroll so the row is on the top left.  This makes it easier to see the
row in question since it's always at the same location, and it shows
more child keybindings.