Fixed some en_GB language usagae
The language usage here was bugging me so I fixed it. Sorry I don't know much about Glade configuration. Looking at the msgid it looks like there is some issues with the source language too. Also I removed an _ on line 2019 I wasn't sure if that was relevant. It didn't appear to be doing anything.