Prevent passing NULL pointer to strcmp()
Without the check, new_name could be set to NULL in the changed
if block and when it `continue`s the loop a few lines down, it
could end up being called again with the newly set NULL pointer.
Not sure it's a real issue but it helps static analyzers navigate
the code better and might prevent future segfaults.