Guard against NULL pointer dereference
The function checked if doc pointer was NULL and later in the
function used it as if it definitively wasn't. If it's not ever
going to be NULL, the new `if (doc == NULL)` check isn't needed but
if it can be, then we should deal with that case. Note: I didn't
use g_return_val_if_fail() because those assertions can be disabled
and I didn't reason too long whether doc pointer could really be NULL
or if it'd just be a programmer's error.
TODO: should check that.