Assert unlikely but possibly NULL pointer dereferences
If the code checks whether pointer is NULL and then goes to use it
anyway regardless, this fools tools like static analyzers into thinking
a NULL pointer could be dereferenced. If the code assert's that won't
happen, the analyzer can proceed happily. Note: I didn't reason too
long about whether the pointers could really be NULL in practice, but
depending on the inputs, it could certainly be asking for problems.