Fix truncation of the data retrieved with --list-documents
Do not truncate the document list to the size of the buffer used for
communication.

Since we now read multiple times, we need to make sure to always send
ETX so a reader don't hang reading the next chunk if the data sent had
exactly the size of the buffer.