Deleted years-old typo
As far as I can tell, the function ```lenlge``` does not exist, and never has. The only places it shows up online are in lists of functions similar to this one (i.e. the source code of other text editing projects, e.g. notepad++ and scite). My guess is that the typo occurred at the same time someone inserted the ```len_trim``` function, which goes between them.