Send selection to (single) custom command
In Geany, one can already send the selection to custom commands, but
those have to be defined beforehand. It's particularly tedious to have
to go to the GUI in order to define a command one will only use
infrequently.

This commit adds the feature that one can send the selection to an
external command that is provided ad-hoc via a dialog input field.

Somewhat related: https://sourceforge.net/p/geany/feature-requests/275/