Open files from selection
In LaTeX a document can be included using \include{filename}.
For this command the extension of the file 'filename' is not necessary.
Therefore, opening the file from a selection did not work. I've
added all the tests that are needed to check if the selected file
exists if the extension of the current file type is added to the
name. Thus, in the example above the file 'filename.tex' is opened
properly.